Move SnackbarDisplay to core-api as pure JVM interface
