+++ core/build.gradle
@@ -34,6 +34,7 @@ dependencies {
   implementation 'io.reactivex.rxjava2:rxandroid:2.1.1'
 
   implementation 'com.jakewharton.timber:timber:4.7.1'
+  implementation 'com.google.android.material:material:1.3.0'
 
   implementation rootProject.ext.retrofit
   implementation rootProject.ext.retrofitGsonConverter
