+++ feature/repo/src/test/java/com/jraska/github/client/repo/RepoDetailViewModelTest.kt
@@ -1,0 +1,5 @@
+package com.jraska.github.client.repo
+
+import org.junit.Assert.*
+
+class RepoDetailViewModelTest
